moved publications to dataclasses module
