+++ mobilizon_reshare/cli/commands/list/list_event.py
@@ -4,8 +4,8 @@ from typing import Iterable, Optional
 import click
 from arrow import Arrow
 
-from mobilizon_reshare.event.event import EventPublicationStatus
-from mobilizon_reshare.event.event import MobilizonEvent
+from mobilizon_reshare.dataclasses.event import EventPublicationStatus
+from mobilizon_reshare.dataclasses.event import MobilizonEvent
 from mobilizon_reshare.event.event_selection_strategies import select_unpublished_events
 from mobilizon_reshare.storage.query.read import (
     get_published_events,
