+++ mobilizon_reshare/dataclasses/publication.py
@@ -1,0 +1,40 @@
+from dataclasses import dataclass
+from typing import List
+from uuid import UUID
+
+from mobilizon_reshare.dataclasses.event import MobilizonEvent
+from mobilizon_reshare.models.publication import Publication
+
+from mobilizon_reshare.publishers.abstract import (
+    AbstractPlatform,
+    AbstractEventFormatter,
+)
+
+
+@dataclass
+class BasePublication:
+    publisher: AbstractPlatform
+    formatter: AbstractEventFormatter
+
+
+@dataclass
+class EventPublication(BasePublication):
+    event: MobilizonEvent
+    id: UUID
+
+    @classmethod
+    def from_orm(cls, model: Publication, event: MobilizonEvent):
+        # imported here to avoid circular dependencies
+        from mobilizon_reshare.publishers.platforms.platform_mapping import (
+            get_publisher_class,
+            get_formatter_class,
+        )
+
+        publisher = get_publisher_class(model.publisher.name)()
+        formatter = get_formatter_class(model.publisher.name)()
+        return cls(publisher, formatter, event, model.id,)
+
+
+@dataclass
+class RecapPublication(BasePublication):
+    events: List[MobilizonEvent]
