+++ mobilizon_reshare/publishers/platforms/twitter.py
@@ -4,7 +4,7 @@ import pkg_resources
 from tweepy import OAuthHandler, API, TweepyException
 from tweepy.models import Status
 
-from mobilizon_reshare.event.event import MobilizonEvent
+from mobilizon_reshare.dataclasses.event import MobilizonEvent
 from mobilizon_reshare.publishers.abstract import (
     AbstractPlatform,
     AbstractEventFormatter,
