+++ mobilizon_reshare/storage/query/write.py
@@ -4,7 +4,7 @@ from typing import Iterable, Optional
 import arrow
 from tortoise.transactions import atomic
 
-from mobilizon_reshare.event.event import MobilizonEvent
+from mobilizon_reshare.dataclasses.event import MobilizonEvent
 from mobilizon_reshare.models.event import Event
 from mobilizon_reshare.models.publication import Publication
 from mobilizon_reshare.models.publisher import Publisher
