+++ tests/publishers/conftest.py
@@ -5,7 +5,7 @@ from uuid import UUID
 import arrow
 import pytest
 
-from mobilizon_reshare.event.event import MobilizonEvent
+from mobilizon_reshare.dataclasses.event import MobilizonEvent
 from mobilizon_reshare.publishers.abstract import (
     AbstractPlatform,
     AbstractEventFormatter,
