+++ tests/storage/test_query.py
@@ -4,9 +4,8 @@ from uuid import UUID
 import arrow
 import pytest
 
-from mobilizon_reshare.event.event import EventPublicationStatus
+from mobilizon_reshare.dataclasses.event import EventPublicationStatus
 from mobilizon_reshare.models.publication import PublicationStatus
-from mobilizon_reshare.publishers.abstract import EventPublication
 from mobilizon_reshare.storage.query.read import (
     get_published_events,
     events_with_status,
@@ -57,7 +56,7 @@ async def test_get_published_events(generate_models):
     ],
 )
 async def test_publications_with_status(
-        status, mobilizon_id, from_date, to_date, expected_result, generate_models,
+    status, mobilizon_id, from_date, to_date, expected_result, generate_models,
 ):
     await generate_models(complete_specification)
     publications = await publications_with_status(
@@ -84,33 +83,33 @@ async def test_event_with_status(generate_models, status, expected_events_count)
     "status, expected_events_count, begin_window, end_window",
     [
         (
-                EventPublicationStatus.COMPLETED,
-                2,
-                arrow.get(today + timedelta(hours=-1)),
-                None,
+            EventPublicationStatus.COMPLETED,
+            2,
+            arrow.get(today + timedelta(hours=-1)),
+            None,
         ),
         (
-                EventPublicationStatus.COMPLETED,
-                1,
-                arrow.get(today + timedelta(hours=1)),
-                None,
+            EventPublicationStatus.COMPLETED,
+            1,
+            arrow.get(today + timedelta(hours=1)),
+            None,
         ),
         (
-                EventPublicationStatus.COMPLETED,
-                1,
-                arrow.get(today + timedelta(hours=-2)),
-                arrow.get(today + timedelta(hours=1)),
+            EventPublicationStatus.COMPLETED,
+            1,
+            arrow.get(today + timedelta(hours=-2)),
+            arrow.get(today + timedelta(hours=1)),
         ),
         (
-                EventPublicationStatus.COMPLETED,
-                0,
-                arrow.get(today + timedelta(hours=-2)),
-                arrow.get(today + timedelta(hours=0)),
+            EventPublicationStatus.COMPLETED,
+            0,
+            arrow.get(today + timedelta(hours=-2)),
+            arrow.get(today + timedelta(hours=0)),
         ),
     ],
 )
 async def test_event_with_status_window(
-        generate_models, status, expected_events_count, begin_window, end_window
+    generate_models, status, expected_events_count, begin_window, end_window
 ):
     await generate_models(complete_specification)
     result = list(
@@ -126,23 +125,23 @@ async def test_event_with_status_window(
     [
         ({"event": 2, "publications": [], "publisher": ["zulip"]}, [event_0, event_1],),
         (
-                {
-                    "event": 3,
-                    "publications": [
-                        {
-                            "event_idx": 1,
-                            "publisher_idx": 0,
-                            "status": PublicationStatus.FAILED,
-                        },
-                        {
-                            "event_idx": 2,
-                            "publisher_idx": 0,
-                            "status": PublicationStatus.COMPLETED,
-                        },
-                    ],
-                    "publisher": ["zulip"],
-                },
-                [event_0],
+            {
+                "event": 3,
+                "publications": [
+                    {
+                        "event_idx": 1,
+                        "publisher_idx": 0,
+                        "status": PublicationStatus.FAILED,
+                    },
+                    {
+                        "event_idx": 2,
+                        "publisher_idx": 0,
+                        "status": PublicationStatus.COMPLETED,
+                    },
+                ],
+                "publisher": ["zulip"],
+            },
+            [event_0],
         ),
         (complete_specification, [event_3],),
     ],
@@ -160,26 +159,26 @@ async def test_events_without_publications(spec, expected_events, generate_model
     [
         ([], {"event": 2, "publications": [], "publisher": ["zulip"]}, event_0, 0,),
         (
-                ["zulip"],
-                {"event": 2, "publications": [], "publisher": ["zulip"]},
-                event_0,
-                1,
+            ["zulip"],
+            {"event": 2, "publications": [], "publisher": ["zulip"]},
+            event_0,
+            1,
         ),
         (
-                ["telegram", "zulip", "mastodon", "facebook"],
-                {
-                    "event": 2,
-                    "publications": [],
-                    "publisher": ["telegram", "zulip", "mastodon", "facebook"],
-                },
-                event_0,
-                4,
+            ["telegram", "zulip", "mastodon", "facebook"],
+            {
+                "event": 2,
+                "publications": [],
+                "publisher": ["telegram", "zulip", "mastodon", "facebook"],
+            },
+            event_0,
+            4,
         ),
     ],
     indirect=["mock_active_publishers"],
 )
 async def test_build_publications(
-        mock_active_publishers, spec, event, n_publications, generate_models
+    mock_active_publishers, spec, event, n_publications, generate_models
 ):
     await generate_models(spec)
